+++ CHANGES.rst
@@ -3,6 +3,8 @@
 
 * implement ``PipFreeze.__bool__()`` (py3) and ``PipFreeze.__nonzero__()`` (py2)
 * implement ``PipFreeze.__iter__()``
+* implement ``Requirement``
+* implement ``Package``
 
 0.0.1
 =====